requirements.txt
# Core LLM and AI Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langgraph>=0.0.23

# API Clients
anthropic>=0.25.0
openai>=1.3.0
google-generativeai>=0.3.0

# Data Processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.28.0

# Optional: Evaluation metrics
ragas>=0.0.80

# Utilities
tiktoken>=0.5.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
